requirements.txt
psycopg2-binary
pyTelegramBotAPI
python-dotenv
matplotlib
numpy
scipy
statsmodels
telebot-calendar